(#148) Moved JsonSchemaValidation from citrus/validation/json to citrus/validation/json/schema
